requirements.txt
flask
requests
python-dotenv
openai
mcp-server
qdrant-client